+++ gradle-conventions-settings/src/main/kotlin/util/projectReport.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package util
+
+import org.gradle.api.Project
+
+// useful for dependencies introspection
+// run ./gradlew htmlDependencyReport
+// Report can normally be found in build/reports/project/dependencies/index.html
+fun Project.configureProjectReport() {
+    allprojects {
+        plugins.apply("project-report")
+    }
+}
